Refactor Navbar and its tests
